# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API Documentation'
  description: "Documentation des API de l'application Les Innovateurs. Cette documentation fournit des informations détaillées sur les différentes API disponibles pour gérer les utilisateurs, les élections, les candidats, les activités, les sondages et d'autres fonctionnalités de l'application."
  license:
    name: 'URL Accueil'
    url: 'https://lesinnovateurs.me'
  version: 1.0.0
paths:
...
  /api/private/user/add-programme-reaction:
    post:
      tags:
        - 'User Authenticated Actions'
      summary: 'Add a reaction to a programme'
      description: 'Add a reaction to a programme'
      operationId: addProgrammeReaction
      requestBody:
        description: 'Programme reaction object that needs to be added to the store'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResultatProgramme'
      responses:
        200:
          description: 'Programme reaction added successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme reaction added successfully'
                type: object
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: Unauthenticated.
                type: object
        422:
          description: "Message d'erreur pour les champs non valides"
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Les données fournies sont invalides'
                  errors:
                    type: object
                type: object
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: 'User login'
      operationId: 535ffeda4b78916efd658b2844ea897a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: password123
              type: object
      responses:
        200:
          description: 'Successful login, returns user and token'
        401:
          description: Unauthorized
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: 'User registration'
      operationId: da8643013c2941f5ed7f8a9c9b4a7f7a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: 'Successful registration'
        400:
          description: 'Bad request'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: 'User logout'
      operationId: 46f8a14c8083abb69eed72768aa08629
      responses:
        200:
          description: 'Successfully logged out'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: 'Refresh authentication token'
      operationId: f3538b6e9c46bcf601a9d194288b03d4
      responses:
        200:
          description: 'Token refreshed'
        401:
          description: Unauthorized
components:
  schemas:
...
    ResultatProgramme:
      title: 'Resultat Programme'
      description: 'Modèle de résultat de programme'
      properties:
        id_programme:
          description: 'ID du programme associé au résultat'
          type: integer
        id_user:
          description: "ID de l'utilisateur lié au résultat"
          type: integer
        avis:
          description: "Avis de l'utilisateur sur le programme"
          type: string
        created_at:
          description: 'Date et heure de création'
          type: string
          format: date-time
          readOnly: 'true'
          nullable: true
        updated_at:
          description: 'Date et heure de mise à jour'
          type: string
          format: date-time
          readOnly: 'true'
          nullable: true
      type: object
...
    User:
      title: User
      description: 'User model'
      properties:
        id:
          description: 'User ID auto incremented'
          type: integer
        nom:
          description: 'User last name'
          type: string
        prenom:
          description: 'User first name'
          type: string
        elector_card:
          description: 'User elector card'
          type: string
        adresse:
          description: 'User address'
          type: string
        numero_cni:
          description: 'User national ID number pas obligatoire'
          type: string
        commune:
          description: 'User commune'
          type: string
        role_id:
          description: 'User role ID (1=admin, 2=candidat, 3=user/electeur,, default=3)'
          type: integer
        phone:
          description: 'User phone number'
          type: string
        email:
          description: 'User email'
          type: string
          format: email
        password:
          description: 'User password'
          type: string
          format: password
        created_at:
          description: 'User creation date and time'
          type: string
          format: date-time
        updated_at:
          description: 'User last update date and time'
          type: string
          format: date-time
        photo_url:
          description: "photo de l'utilisateur pas obligatoire"
          type: file
      type: object
...